Move CustomGlyphs.ts into core
